Getting a snapshot should be nullable
We need to ensure that when successfully fetching a document that
doesn't exist, rather than getting back a document, you get back nil. We
were simply returning the value before checking if it was valid to do
so. This adds the check on whether or not it's a valid document before
returning it back to the caller.