Add support for calling the two-arg variant of runTransaction
When given a first argument closure that returns a non-optional type,
the compiler would give up trying to match against the two-arg variant
and instead complain about the first argument missing a `with:` label
and not being a `TransactionOption` type.

A fix for this is to provide optional and non-optional variants of
the two arg override.