Expose firestore settings (#49)
Switches to wrapping the C++ `Firestore` object with a Swift class so that the `Firestore` object takes on reference semantics. This allows for code like `Firestore.firestore().settings = ...`

Renames `Settings` to `FirestoreSettings`, also providing a wrapper. This time using value semantics. But the wrapper helps hide the C++ methods that are named differently than the Obj C API we are trying to emmulate.

Expose `isPersistenceEnabled` setting.