FirebaseAndroid: introduce the new JNI module for Android support
In order to interact with the C++ Firebase SDK, we need to have access
to the JVM and the JavaEnvironment. Introduce a JNI layer which allows
us to capture this information when loaded. Additionally, register a
`company.thebrowser.Native.RegisterActivity` method to allow us access
to the `Activity`. This is required to be called before the Firebase
layer is accessed by the application.