FirebaseAndroid: fix & remove nullability annotations
The nullability annotations seem to cause warnings due to
`-Wnullability-completeness`. Correctly annotating the types as
`_Nullable JavaVM * _Nonnull` seems to not work across Swift and C/C++.
Rather than try to force this, simply avoid the nullability annotations.

Clear up the method registration which was never being invoked due to an
early return in the success case.