[WPP-2267] Avoid HttpsCallableReference copy constructor (#67)
Unfortunately `HttpsCallableReference` does not use internal reference counting, and as a result, we need to avoid the copy-constructor for `HttpsCallableReference`. Otherwise, if Swift creates a copy of the object and deletes that copy, it will result in any pending `Call` being interrupted or triggering memory corruption in the case that `Call` has not completed. To avoid this and to prevent Swift from seeing the copy constructor, wrap with a `std::shared_ptr`.