Firestore: make `FirebaseFirestore` build on Android
The android build seems to not need the absl, upb, or SSL/crypto
libraries. Extract the absl, upb, and their dependencies into a separate
Windows conditional group. Use the same pattern as the other libraries
for SSL/crypto where we use a generator expression. Add an additional
cast for the error code type difference.